# Reduce # of allocated objects