… src/WinRT.Runtime/Projections/IReadOnlyList.net5.cs