- Replace delegates with function pointers where possible
- Remove unnecessary creation of IMarshal object to optimize CreateObject scenario
- Cache subscribe / unsubscribe actions to avoid creating them each time
- Fix some warnings

Before

|                                   Method |          Mean |         Error |        StdDev |   Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|----------------------------------------- |--------------:|--------------:|--------------:|--------:|-------:|-------:|----------:|
| AddAndInvokeMultipleIntEventsToSameEventSource | 19,042.9 ns | 365.40 ns | 406.14 ns | 0.4883 | 0.2136 | 0.0610 |   2,649 B |
| CreateAndIterateList |  10,524.33 ns |    188.582 ns |    251.752 ns |  0.7477 | 0.3662 |      - |   3,144 B |

After

|                                   Method |          Mean |         Error |        StdDev |   Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|----------------------------------------- |--------------:|--------------:|--------------:|--------:|-------:|-------:|----------:|
| AddAndInvokeMultipleIntEventsToSameEventSource | 18,977.4 ns | 363.47 ns | 432.69 ns | 0.4272 | 0.1526 | 0.0305 |   2,442 B |
|                     CreateAndIterateList |   9,119.60 ns |    174.545 ns |    171.427 ns |  0.6409 | 0.3204 |      - |   2,697 B |